requirements.txt
tqdm
nodriver
requests
aiohttp
beautifulsoup4
pillow